+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [1.1.4] - 2021-12-14
 
 ### Fixed
-* Fix JS loading issue by chaning the inject priority
+* Fix JS loading issue by changing the inject priority
 
 ## [1.1.3] - 2021-12-08